moved out two function to utils
